Fixed spelling of the captured error message.
Changed "File not Found" to "File not found" error as per the Dropbox API error
messages. The new code only allows entries to be posted if the old file was
successfully downloaded or of it never existed to avoid accidentally clobbering
existing entries. Miss-catching the error caused new posts for new months to be
suppressed.

Fixes #6